Compute the Average Causal Effect (ACE) for a given source-sink pair

Compute total effects as ACEs of source variables X (i.e., incoming connectivity = 0) on sink variables Y (i.e., outgoing connectivity = 0), in a directed graph. The ACE will be estimated as the path coefficient of X (i.e., theta) in the linear equation Y ~ X + Z. Z is defined as the adjustment (or conditioning) set of Y over X, applying an "optimal" valid set (O-set), with the smallest asymptotic variance. Standard errors (SE), for each ACE, are computed following the lavaan standard procedure or a bootstrap-based procedure (see boot for details).

SEMace(
  graph,
  data,
  group = NULL,
  method = "none",
  alpha = 0.05,
  boot = NULL,
  ...
)

Arguments

graph

An igraph object.

data

A matrix or data.frame. Rows correspond to subjects, and columns to graph nodes (variables).

group

A binary vector. This vector must be as long as the number of subjects. Each vector element must be 1 for cases and 0 for control subjects. If group = NULL (default), group influence will not be considered.

method

Multiple testing correction method. One of the values available in p.adjust. By default, method = "none" (i.e., no multiple test correction).

alpha

Significance level for ACE selection (by default, alpha = 0.05).

boot

The number of bootstrap samplings enabling bootstrap computation of ACE standard errors. If NULL (default), bootstrap is disabled.

...

Currently ignored.

Value

A data.frame of ACE estimates between network sources and sinks.

References

Witte J, Henckel L, Maathuis MH, Didelez V (2020). On efficient adjustment in causal graphs. arXiv:2002.06825 [math.ST]. <https://arxiv.org/abs/2002.06825/>

Author

Mario Grassi mario.grassi@unipv.it

Examples


# ACE estimation, without group (default)
ace <- SEMace(graph = sachs$graph, data = log(sachs$pkc))
#> WARNING: input graph is not acyclic!
#> 		 Applying graph -> DAG conversion.
#> DAG conversion: TRUE
#> 
#> Frequency distribution of path length from X to Y :
#> 1 3 4 
#> 5 2 1 
#> 
#>  ACE 1 of 8 ACE 2 of 8 ACE 3 of 8 ACE 4 of 8 ACE 5 of 8 ACE 6 of 8 ACE 7 of 8 ACE 8 of 8
print(ace)
#>   sink op source    est    se      z pvalue ci.lower ci.upper
#> 1  Jnk <-    PKA  0.093 0.024  3.938  0.000    0.047    0.139
#> 3  Akt <-    PKA  0.546 0.020 27.439  0.000    0.507    0.585
#> 4  Erk <-    PKA  0.458 0.021 21.582  0.000    0.416    0.499
#> 5  Jnk <-   PIP3 -0.077 0.024 -3.250  0.001   -0.124   -0.031
#> 7  Akt <-   PIP3 -0.065 0.020 -3.249  0.001   -0.104   -0.026
#> 8  Erk <-   PIP3 -0.063 0.021 -2.988  0.003   -0.105   -0.022

# ACE estimation, with group perturbation and multiple test correction
ace2 <- SEMace(graph = sachs$graph, data = log(sachs$pkc),
               group = sachs$group,
               method = "BH", alpha = 0.05)
#> WARNING: input graph is not acyclic!
#> 		 Applying graph -> DAG conversion.
#> DAG conversion: TRUE
#> 
#> Frequency distribution of path length from X to Y :
#> 1 3 4 
#> 5 2 1 
#> 
#>  ACE 1 of 8 ACE 2 of 8 ACE 3 of 8 ACE 4 of 8 ACE 5 of 8 ACE 6 of 8 ACE 7 of 8 ACE 8 of 8
print(ace2)
#>   sink op source d_est  d_se   d_z pvalue d_lower d_upper
#> 3  Akt <-    PKA 0.341 0.039 8.722      0   0.265   0.418
#> 4  Erk <-    PKA 0.242 0.042 5.719      0   0.159   0.326