Compute the Average Causal Effect (ACE) for a given source-sink pair

Compute total effects as ACEs of variables X on variables Y in a directed acyclic graph (DAG). The ACE will be estimated as the path coefficient of X (i.e., theta) in the linear equation Y ~ X + Z. The set Z is defined as the adjustment (or conditioning) set of Y over X, applying various adjustement sets. Standard errors (SE), for each ACE, are computed following the lm standard procedure or a bootstrap-based procedure (see boot for details).

SEMace(
  graph,
  data,
  group = NULL,
  type = "parents",
  effect = "all",
  method = "BH",
  alpha = 0.05,
  boot = NULL,
  ...
)

Arguments

graph

An igraph object.

data

A matrix or data.frame. Rows correspond to subjects, and columns to graph nodes (variables).

group

A binary vector. This vector must be as long as the number of subjects. Each vector element must be 1 for cases and 0 for control subjects. If group = NULL (default), group influence will not be considered.

type

character Conditioning set Z. If "parents" (default) the Pearl's back-door set (Pearl, 1998), "minimal" the dagitty minimal set (Perkovic et al, 2018), or "optimal" the O-set with the smallest asymptotic variance (Witte et al, 2020) are computed.

effect

character X to Y effect. If "all" (default) all effects from X to Y, "source2sink" only effects from source X to sink Y, or "direct" only direct effects from X to Y are computed.

method

Multiple testing correction method. One of the values available in p.adjust. By default, method = "BH" (i.e., FDR multiple test correction).

alpha

Significance level for ACE selection (by default, alpha = 0.05).

boot

The number of bootstrap samplings enabling bootstrap computation of ACE standard errors. If NULL (default), bootstrap is disabled.

...

Currently ignored.

Value

A data.frame of ACE estimates between network sources and sinks.

References

Pearl J (1998). Graphs, Causality, and Structural Equation Models. Sociological Methods & Research, 27(2):226-284. <https://doi.org/10.1177/0049124198027002004>

Perkovic E, Textor J, Kalisch M, Maathuis MH (2018). Complete graphical characterization and construction of adjustment sets in Markov equivalence classes of ancestral graphs. Journal of Machine Learning Research, 18:1-62. <http://jmlr.org/papers/v18/16-319.html>

Witte J, Henckel L, Maathuis MH, Didelez V (2020). On efficient adjustment in causal graphs. Journal of Machine Learning Research, 21:1-45. <http://jmlr.org/papers/v21/20-175.htm>

Author

Mario Grassi mario.grassi@unipv.it

Examples


# ACE estimation, without group
ace1 <- SEMace(graph = sachs$graph, data = log(sachs$pkc),
               group = NULL, type="parents", effect="all",
               method = "BH", alpha = 0.05)
#> 
#>  Frequency distribution of path length from X to Y :
#>  1  2  3 
#> 20 18  3 
#> 
#>  ACE= 1 of 41 ACE= 2 of 41 ACE= 3 of 41 ACE= 4 of 41 ACE= 5 of 41 ACE= 6 of 41 ACE= 7 of 41 ACE= 8 of 41 ACE= 9 of 41 ACE= 10 of 41 ACE= 11 of 41 ACE= 12 of 41 ACE= 13 of 41 ACE= 14 of 41 ACE= 15 of 41 ACE= 16 of 41 ACE= 17 of 41 ACE= 18 of 41 ACE= 19 of 41 ACE= 20 of 41 ACE= 21 of 41 ACE= 22 of 41 ACE= 23 of 41 ACE= 24 of 41 ACE= 25 of 41 ACE= 26 of 41 ACE= 27 of 41 ACE= 28 of 41 ACE= 29 of 41 ACE= 30 of 41 ACE= 31 of 41 ACE= 32 of 41 ACE= 33 of 41 ACE= 34 of 41 ACE= 35 of 41 ACE= 36 of 41 ACE= 37 of 41 ACE= 38 of 41 ACE= 39 of 41 ACE= 40 of 41 ACE= 41 of 41
print(ace1)
#>    pathL sink op source    est    se      z pvalue ci.lower ci.upper
#> 1      3  PKC <-    PKA  0.062 0.024  2.610  0.009    0.016    0.109
#> 2      1 PIP3 <-    PKA  0.000 0.000  4.869  0.000    0.000    0.000
#> 3      1  Mek <-    PKA -0.067 0.024 -2.834  0.005   -0.114   -0.021
#> 4      1  Raf <-    PKA -0.111 0.023 -4.722  0.000   -0.157   -0.065
#> 5      2 PIP2 <-    PKA  0.050 0.020  2.474  0.013    0.010    0.089
#> 6      2 Plcg <-    PKA  0.084 0.023  3.591  0.000    0.038    0.130
#> 7      1  Jnk <-    PKA  0.093 0.024  3.907  0.000    0.046    0.139
#> 9      1  Akt <-    PKA  0.546 0.020 27.439  0.000    0.507    0.585
#> 10     1  Erk <-    PKA  0.451 0.021 21.257  0.000    0.410    0.493
#> 13     1  Jnk <-    PKC  0.133 0.024  5.640  0.000    0.087    0.179
#> 14     1  P38 <-    PKC  0.645 0.018 35.449  0.000    0.609    0.680
#> 16     1  PKA <-   PIP3  0.000 0.000  4.628  0.000    0.000    0.000
#> 18     2  Mek <-   PIP3  0.085 0.024  3.560  0.000    0.038    0.131
#> 19     2  Raf <-   PIP3  0.155 0.023  6.600  0.000    0.109    0.201
#> 20     1 PIP2 <-   PIP3  0.541 0.020 26.839  0.000    0.502    0.581
#> 21     1 Plcg <-   PIP3  0.216 0.023  9.245  0.000    0.170    0.261
#> 22     2  Jnk <-   PIP3 -0.077 0.024 -3.250  0.001   -0.124   -0.031
#> 24     1  Akt <-   PIP3 -0.065 0.020 -3.249  0.001   -0.104   -0.026
#> 25     2  Erk <-   PIP3 -0.063 0.021 -2.988  0.003   -0.105   -0.022
#> 26     1  Erk <-    Mek  0.063 0.029  2.176  0.030    0.006    0.119
#> 27     1  Mek <-    Raf  0.683 0.018 38.898  0.000    0.649    0.718
#> 28     2  Erk <-    Raf -0.081 0.021 -3.784  0.000   -0.122   -0.039
#> 31     2  Raf <-   PIP2 -0.088 0.029 -3.089  0.002   -0.144   -0.032
#> 36     2  Mek <-   Plcg -0.065 0.024 -2.680  0.007   -0.112   -0.017
#> 37     2  Raf <-   Plcg -0.100 0.024 -4.191  0.000   -0.147   -0.053
#> 38     1 PIP2 <-   Plcg  0.227 0.020 11.435  0.000    0.188    0.265
#> 39     2  Jnk <-   Plcg  0.111 0.024  4.620  0.000    0.064    0.159
#> 41     3  Erk <-   Plcg  0.096 0.024  3.968  0.000    0.048    0.143

# ACE with group perturbation
ace2 <- SEMace(graph = sachs$graph, data = log(sachs$pkc),
               group = sachs$group, type="optimal", effect="direct",
               method = "none", alpha = 0.05)
#> 
#> WARNING: input graph is not acyclic!
#>  Applying graph -> DAG conversion.
#> DAG conversion: TRUE
#> 
#>  Frequency distribution of path length from X to Y :
#>  1 
#> 18 
#> 
#>  ACE= 1 of 18 ACE= 2 of 18 ACE= 3 of 18 ACE= 4 of 18 ACE= 5 of 18 ACE= 6 of 18 ACE= 7 of 18 ACE= 8 of 18 ACE= 9 of 18 ACE= 10 of 18 ACE= 11 of 18 ACE= 12 of 18 ACE= 13 of 18 ACE= 14 of 18 ACE= 15 of 18 ACE= 16 of 18 ACE= 17 of 18 ACE= 18 of 18
print(ace2)
#>    pathL sink op source d_est  d_se    d_z pvalue d_lower d_upper
#> 5      1  Akt <-    PKA 0.341 0.039  8.722  0.000   0.265   0.418
#> 6      1  Erk <-    PKA 0.242 0.042  5.719  0.000   0.159   0.326
#> 9      1  Jnk <-    PKC 0.645 0.044 14.523  0.000   0.558   0.732
#> 10     1  P38 <-    PKC 0.347 0.036  9.739  0.000   0.277   0.417
#> 11     1 PIP2 <-   PIP3 0.351 0.040  8.715  0.000   0.272   0.430
#> 12     1 Plcg <-   PIP3 0.309 0.047  6.529  0.000   0.216   0.402
#> 13     1  Akt <-   PIP3 0.079 0.040  1.994  0.046   0.001   0.158
#> 18     1 PIP2 <-   Plcg 0.375 0.039  9.715  0.000   0.299   0.450