SEM-based gene set analysis

Gene Set Analysis (GSA) via self-contained test for group effect on signaling (directed) pathways as SEM, evaluating overall pathway perturbation, perturbation emission from source nodes, and perturbation accumulation on target nodes. Approximate randomization test P-values of specific node and aggregated group effects will be computed. For directed graphs, they include: the sum of group effects adjusted by residual variances (D), the sum of the tagret nodes perturbation (i.e., group effect) accumulation from source nodes (A), and the sum of the source nodes perturbation emission towards target nodes (E). For undirected graphs, the sum of group effects, adjusted by residual variances (D), will be estimated.

SEMgsa(g = list(), data, group, method = "BH", alpha = 0.05, n_rep = 1000, ...)

Arguments

g
A list of pathways to be tested.

data
A matrix or data.frame. Rows correspond to subjects, and columns to graph nodes (variables).

group
A binary vector. This vector must be as long as the number of subjects. Each vector element must be 1 for cases and 0 for control subjects.

method
Multiple testing correction method. One of the values available in p.adjust. By default, method is set to "BH" (i.e., Benjamini-Hochberg correction).

alpha
Gene set test significance level. Alpha is set to 0.05 by default.

n_rep
Number of randomization replicates (default = 1000).

...
Currently ignored.

Value

A list of 2 objects:

"gsa", A data.frame reporting the following information for each pathway in the input list:

"N.nodes", pathway size (number of nodes);

"N.DRNs", number of differential regulated genes (DRNs) within the pathway, after multiple test correction with Benjamini-Hochberg method;

"pD", significance of the sum of group effects, adjusted by the residual variance;

"pA", significance of the sum of tagret nodes perturbation (i.e., group effect) accumulation from source nodes;

"pE", significance of the sum of source nodes perturbation (i.e., group effect) emission towards target nodes;

"pvalue", Fisher's combined P-value of pD, pA, and pE.

"DRN", a list with DRNs names per pathways.

Author

Mario Grassi mario.grassi@unipv.it

Examples


# \dontrun{

# Install data examples, reference networks, and pathways
#devtools::install_github("fernandoPalluzzi/SEMdata")
library(SEMdata)

#> 
#> Attaching package: 'SEMdata'

#> The following objects are masked from 'package:SEMgraph':
#> 
#>     alsData, kegg

library(huge)

als.npn <- huge.npn(alsData$exprs)

#> Conducting the nonparanormal (npn) transformation via shrunkun ECDF....done.


#selection of FTD pathways from KEGG site:
# https://www.genome.jp/dbget-bin/www_bget?ds:H00078

paths.name<- c(
 "MAPK signaling pathway",
 "Protein processing in endoplasmic reticulum",
 "Endocytosis",
 "Wnt signaling pathway",
 "Notch signaling pathway",
 "Neurotrophin signaling pathway",
 "Amyotrophic lateral sclerosis (ALS)")

I <- which(names(kegg.pathways) %in% paths.name)
GSA <- SEMgsa(kegg.pathways[I], als.npn, alsData$group,
              method = "bonferroni", alpha = 0.05,
              n_rep = 1000)

#> # 1: MAPK signaling pathway
#> # 2: Protein processing in endoplasmic reticulum
#> # 3: Endocytosis
#> # 4: Wnt signaling pathway
#> # 5: Notch signaling pathway
#> # 6: Neurotrophin signaling pathway
#> # 7: Amyotrophic lateral sclerosis (ALS)

GSA$gsa

#>                                             N.nodes N.DRNs       pD       pA
#> MAPK signaling pathway                          292     47 0.915807 0.010656
#> Protein processing in endoplasmic reticulum      26     10 0.877232 0.000002
#> Endocytosis                                      29      7 0.000001 0.000003
#> Wnt signaling pathway                           157     33 0.017545 0.923278
#> Notch signaling pathway                          53     14 0.000049 0.000429
#> Neurotrophin signaling pathway                  117     32 0.003745 0.006109
#> Amyotrophic lateral sclerosis (ALS)              32      7 0.013752 0.000541
#>                                                   pE   pvalue
#> MAPK signaling pathway                      0.410353 0.087129
#> Protein processing in endoplasmic reticulum 0.052767 0.000014
#> Endocytosis                                 0.478188 0.000000
#> Wnt signaling pathway                       0.774848 0.187785
#> Notch signaling pathway                     0.000000 0.000000
#> Neurotrophin signaling pathway              0.208317 0.000421
#> Amyotrophic lateral sclerosis (ALS)         0.287008 0.000212

GSA$DRN

#> $`MAPK signaling pathway`
#>  [1] "207"       "208"       "774"       "775"       "776"       "9254"     
#>  [7] "836"       "1398"      "1436"      "1946"      "8822"      "2885"     
#> [13] "3480"      "8517"      "8681"      "4254"      "4296"      "7786"     
#> [19] "4215"      "9064"      "8491"      "1432"      "5599"      "23162"    
#> [25] "5601"      "2122"      "8569"      "4763"      "4772"      "4775"     
#> [31] "56034"     "5228"      "100137049" "5494"      "5532"      "5534"     
#> [37] "5567"      "84867"     "5879"      "5881"      "5906"      "5921"     
#> [43] "5922"      "9252"      "6654"      "23118"     "57551"    
#> 
#> $`Protein processing in endoplasmic reticulum`
#>  [1] "578"   "1649"  "27102" "2081"  "5602"  "5599"  "5601"  "4780"  "23645"
#> [10] "7466" 
#> 
#> $Endocytosis
#> [1] "867"   "2060"  "83737" "4193"  "51100" "56904" "11059"
#> 
#> $`Wnt signaling pathway`
#>  [1] "10297" "816"   "894"   "896"   "57680" "1452"  "80319" "23002" "1856" 
#> [10] "23291" "7976"  "55366" "8549"  "6885"  "4609"  "4772"  "4776"  "64840"
#> [19] "5532"  "5534"  "5567"  "5879"  "5881"  "9978"  "387"   "9475"  "84870"
#> [28] "8607"  "59343" "6477"  "6907"  "57216" "80326"
#> 
#> $`Notch signaling pathway`
#>  [1] "9541"  "54567" "1856"  "3066"  "182"   "3714"  "2648"  "8850"  "9794" 
#> [10] "55534" "4854"  "5664"  "3516"  "5986" 
#> 
#> $`Neurotrophin signaling pathway`
#>  [1] "207"   "10000" "91860" "814"   "1399"  "2309"  "10818" "11213" "3667" 
#> [10] "3845"  "5604"  "4215"  "4217"  "5602"  "5600"  "5599"  "5601"  "4793" 
#> [19] "4893"  "4916"  "5296"  "5335"  "5336"  "5781"  "5879"  "5906"  "387"  
#> [28] "8767"  "9252"  "6272"  "6654"  "7531" 
#> 
#> $`Amyotrophic lateral sclerosis (ALS)`
#> [1] "836"  "5608" "5600" "4747" "4741" "5532" "5534"
#> 


# }