Conditional Independence (CI) local tests of an acyclic graph

P-values of one minimal testable implication (with the smallest possible conditioning set) is returned per missing edge given an acyclic graph (DAG or BAP) using the function impliedConditionalIndependencies plus the function localTests from package dagitty. Without assuming any particular dependence structure, the p-values of every CI test, in a DAG (BAP), is then combined using the Bonferroni’s statistic in an overall test of the fitted model, B = K*min(p1,...,pK), as reviewed in Vovk & Wang (2020).

localCI.test(graph, data, bap = FALSE, limit = 10000, verbose = TRUE, ...)

Arguments

graph

A directed graph as an igraph object.

data

A data matrix with subjects as rows and variables as columns.

bap

If TRUE, the input graph is trasformend in a BAP, if FALSE (defult) the input graph is reduced in a DAG.

limit

An integer value corresponding to the number of missing edges of the extracted acyclic graph. Beyond this limit, switch to Shipley's C-test (Shipley 2000) is enabled to reduce the computational burden. By default, limit = 10000.

verbose

If TRUE, LocalCI results will be showed to screen (default = TRUE).

...

Currently ignored.

Value

A list of three objects: (i) the DAG used to perform the localCI test (ii) the list of all d-separation tests over missing edges in the input graph and (iii) the overall Bonferroni's P-value.

References

Vovk V, Wang R (2020). Combining p-values via averaging. Biometrika 107(4): 791-808. <https://doi.org/10.1093/biomet/asaa027>

Shipley B (2000). A new inferential test for path models based on DAGs. Structural Equation Modeling, 7(2): 206-218. <https://doi.org/10.1207/S15328007SEM0702_4>

Author

Mario Grassi mario.grassi@unipv.it

Examples


library(huge)
als.npn <- huge.npn(alsData$exprs)
#> Conducting the nonparanormal (npn) transformation via shrunkun ECDF....done.

sem <- SEMrun(alsData$graph, als.npn)
#> NLMINB solver ended normally after 8 iterations 
#> 
#> deviance/df: 10.92479  srmr: 0.2858233 
#> 
B_test <- localCI.test(sem$graph, als.npn, verbose = TRUE)
#> d-separation test (minimal set) of 420 edges...
#>     B_test df B_pvalue
#> 1 376.6275  1        0