Graph properties summary and graph decomposition

Produces a summary of network properties and returns graph components (ordered by decreasing size), without self-loops.

properties(graph, data = NULL, ...)

Arguments

graph

Input network as an igraph object.

data

An optional data matrix (default data = NULL) whith rows corresponding to subjects, and columns to graph nodes (variables). Nodes will be mapped onto variable names.

...

Currently ignored.

Value

List of graph components, ordered by decreasing size (the first component is the giant one), without self-loops.

Author

Mario Grassi mario.grassi@unipv.it

Examples


# Extract the "Type II diabetes mellitus" pathway:
g <- kegg.pathways[["Type II diabetes mellitus"]]
summary(g)
#> IGRAPH 105832d DNW- 46 124 -- 
#> + attr: name (v/c), weight (e/n)
properties(g)
#> Frequency distribution of graph components
#> 
#>   n.nodes n.graphs
#> 1       7        1
#> 2      38        1
#> 
#> Percent of vertices in the giant component: 82.6 %
#> 
#>   is.simple      is.dag is.directed is.weighted 
#>        TRUE       FALSE        TRUE        TRUE 
#> 
#> which.mutual.FALSE 
#>                114 
#> [[1]]
#> IGRAPH 93a670e DNW- 38 114 -- 
#> + attr: name (v/c), weight (e/n)
#> + edges from 93a670e (vertex names):
#>  [1] 5290  ->6517 5290  ->5590 5290  ->2475 5291  ->6517 5291  ->5590
#>  [6] 5291  ->2475 5293  ->6517 5293  ->5590 5293  ->2475 5295  ->6517
#> [11] 5295  ->5590 5295  ->2475 5296  ->6517 5296  ->5590 5296  ->2475
#> [16] 8503  ->6517 8503  ->5590 8503  ->2475 3667  ->5290 3667  ->5291
#> [21] 3667  ->5293 3667  ->5295 3667  ->5296 3667  ->8503 3643  ->3667
#> [26] 3643  ->5594 3643  ->5595 3643  ->8471 3643  ->8660 3630  ->3643
#> [31] 122809->3667 122809->3643 122809->8471 122809->8660 8651  ->3667
#> [36] 8651  ->3643 8651  ->8471 8651  ->8660 8835  ->3667 8835  ->3643
#> + ... omitted several edges
#> 
#> [[2]]
#> IGRAPH 93a670e DNW- 7 10 -- 
#> + attr: name (v/c), weight (e/n)
#> + edges from 93a670e (vertex names):
#>  [1] 2645 ->5313 2645 ->5315 3098 ->5313 3098 ->5315 3099 ->5313 3099 ->5315
#>  [7] 3101 ->5313 3101 ->5315 80201->5313 80201->5315
#>