Graph weighting methods

Add data-driven edge and node weights to the input graph.

weightGraph(
  graph,
  data,
  group = NULL,
  method = "r2z",
  seed = "none",
  limit = 10000,
  ...
)

Arguments

graph

An igraph object.

data

A matrix or data.frame. Rows correspond to subjects, and columns to graph nodes.

group

Binary vector. This vector must be as long as the number of subjects. Each vector element must be 1 for cases and 0 for control subjects. By default, group = NULL.

method

Edge weighting method. It can be one of the following:

"r2z", weight edges of a graph using Fisher's r-to-z transform to test the group difference between correlation coefficients of pairs of interacting nodes (Fisher, 1915).

"sem", edge weights are defined by a SEM model that implies testing the group effect simultaneously on the j-th source node and the k-th sink node. A new parameter w is defined as the weighted sum of the total effect of the group on source and sink nodes, adjusted by node degree centrality, and edge weights correspond to the sign and P-value of the z-test = w/SE(w). Not available if group == NULL.

"cov", edge weights are defined by a new parameter w combining the group effect on the source node (mean group difference, adjusted by source degree centrality), the sink node (mean group difference, adjusted by sink degree centrality), and the source-sink interaction (correlation difference). Edge weights correspond to the sign and P-value of the z-test = w/SE(w) of the combined difference of the group over source node, sink node, and their connection. Not available if group == NULL.

seed

A vector of three cutoffs. By default, seed = "none" and seed calculation is disabled. Suggested cutoff values are seed = c(0.05, 0.5, 0.9). If these cutoffs are defined, seed search is enabled. Nodes can be labeled as either seeds (node weight = 1) or non-seeds (node weight = 0), according to three alternative importance criteria: perturbed group effect, prototype clustering, and closeness node index. The first cutoff is the significance level of the group effect over graph nodes. The second is a threshold corresponding to the prototype clustering distance measure (= 1 - abs(correlation)) cutoff. The third one is the closeness percentile, nodes having closeness greater than the q-th percentile are labeled as seeds. If the seed argument is enabled, the output graph will have three new binary (1: seed, 0: non-seed) vertex attributes:

"pvlm", adjusted FDR p-value < alpha seeds from simple linear regression y ~ x (i.e., lm(node ~ group));

"proto", prototype seeds derived from protoclust;

"qi", nodes with closeness greater than the q-th percentile.

limit

An integer value corresponding to the number of graph edges. Beyond this limit, multicore computation is enabled to reduce the computational burden. By default, limit = 10000.

...

Currently ignored.

Value

A weighted graph, as an igraph object.

References

Palluzzi F, Grassi M (2021). SEMgraph: An R Package for Causal Network Analysis of High-Throughput Data with Structural Equation Models. <arXiv:2103.08332>

Author

Mario Grassi mario.grassi@unipv.it

Examples


# Graph weighting
G <- weightGraph(graph = sachs$graph,
                 data = log(sachs$pkc),
                 group = sachs$group,
                 method = "r2z",
                 seed = c(0.05, 0.5, 0.5))

# New edge attributes
E(G)$pv
#>  [1] 4.401098e-01 5.167191e-61 4.742076e-01 5.848183e-07 1.003647e-09
#>  [6] 9.642826e-01 2.910306e-03 6.797804e-01 8.370494e-01 2.164419e-01
#> [11] 5.465880e-01 5.528435e-01 8.129283e-02 8.879150e-02 4.886605e-01
#> [16] 2.352088e-23 1.495446e-36 3.326762e-14 2.661577e-01 2.661577e-01
E(G)$zsign
#>  [1] 0 1 0 1 1 0 1 0 0 0 0 0 0 0 0 1 1 1 0 0

# New nodes attributes (1: seed, 0: non-seed)
V(G)$pvlm; table(V(G)$pvlm)
#>  [1] 1 1 1 0 1 1 1 1 1 0 1
#> 
#> 0 1 
#> 2 9 
V(G)$proto; table(V(G)$proto)
#>  [1] 0 1 0 0 1 1 1 1 0 1 0
#> 
#> 0 1 
#> 5 6 
V(G)$qi; table(V(G)$qi)
#>  [1] 1 1 1 1 1 0 0 0 0 0 0
#> 
#> 0 1 
#> 6 5 

# Reduced graph (using highest closeness nodes)
R <- induced_subgraph(G, vids = V(G)$name[V(G)$qi == 1])
R <- properties(R)[[1]]
#> Frequency distribution of graph components
#> 
#>   n.nodes n.graphs
#> 1       5        1
#> 
#> Percent of vertices in the giant component: 100 %
#> 
#>   is.simple      is.dag is.directed is.weighted 
#>        TRUE       FALSE        TRUE       FALSE 
#> 
#> which.mutual.FALSE  which.mutual.TRUE 
#>                  5                  2